consolidate twitter utils
